+++ DBADash/DBCollector.cs
@@ -166,7 +166,7 @@ namespace DBADash
             return new ResiliencePipelineBuilder()
                 .AddRetry(new RetryStrategyOptions
                 {
-                    ShouldHandle = new PredicateBuilder().Handle<Exception>(ShouldRetry),
+                    ShouldHandle = new PredicateBuilder().Handle<Exception>(Utility.ShouldRetry),
                     DelayGenerator = static args =>
                     {
                         var delay = args.AttemptNumber switch
@@ -286,32 +286,7 @@ namespace DBADash
             LogInternalPerformanceCounter("DBADash", "Collection Duration (ms)", currentCollection, Convert.ToDecimal(swatch.Elapsed.TotalMilliseconds));
         }
 
-        private static readonly FrozenSet<int> ExcludedErrorCodes =
-        [
-            -2, // retryPolicy excludes query timeout #581
-            218, // Could not find the type '%.*ls'. Either it does not exist or you do not have the necessary permission.
-            219, // The type '%.*ls' already exists, or you do not have permission to create it.
-            229, // The %ls permission was denied on the object '%.*ls', database '%.*ls', schema '%.*ls'.
-            230, //	The %ls permission was denied on the column '%.*ls' of the object '%.*ls', database '%.*ls', schema '%.*ls'.
-            245,   // Conversion failed when converting the %ls value '%.*ls' to data type %ls.
-            262, // %ls permission denied in database '%.*ls'.
-            297, //The user does not have permission to perform this action.
-            300, // %ls permission was denied on object '%.*ls', database '%.*ls'.
-            349,  // The procedure "%.*ls" has no parameter named "%.*ls".
-            500,  // Trying to pass a table-valued parameter with %d column(s) where the corresponding user-defined table type requires %d column(s).
-            2812, // Could not find stored procedure '%.*ls'.
-            6335, // XML data type instance has too many levels of nested nodes. Maximum allowed depth is %d levels.
-            8134, // Divide by zero error encountered.
-        ];
-
-        public static bool ShouldRetry(Exception ex)
-        {
-            if (ex is SqlException sqlEx)
-            {
-                return !ExcludedErrorCodes.Contains(sqlEx.Number) && sqlEx.Message != "Max databases exceeded for Table Size collection";
-            }
-            return true;
-        }
+ 
 
         private async Task StartupAsync()
         {
@@ -344,7 +319,7 @@ namespace DBADash
                     }
                     catch (Exception ex)
                     {
-                        if (ShouldRetry(ex) && !DisableRetry)
+                        if (Utility.ShouldRetry(ex) && !DisableRetry)
                         {
                             Log.Warning(ex, "Error collecting {collection}", collection);
                         }
